Fix Bug 870486: Add snippet templates.
Adds a model SnippetTemplate, which contains a jinja2 template that is
used to render a snippet. Info about the variables available in the
template is stored as SnippetTemplateVariables.

The Snippet admin now presents a widget that lets admins choose which
template to use, and shows them fields for filling in the variables
from the template. When rendered, the snippets render the template,
passing in the data entered in the admin to generate the output.