Temporarily disable the multiple database router.
dumpdata, which is used by django-smuggler, does a check to see if the
model you are dumping is allowed for syncing by the current database.
However, django-smuggler passes an empty string for the name of the
current database, which causes PinningMasterSlaveRouter to return 
False for this check, as it only allows syncing on the master 
database.

Since we're currently on hold with regards to using multiple 
databases, we can remove PinningMasterSlaveRouter and add it back in
with a solution once we solve the multiple database issues.