Don't show disabled snippets in public view and fix admin bulk save.
Bulk save in the admin was broken because it uses save_model, which
assumed there was a `locales` field in the form. However, the form
used for saving multiple snippets at once doesn't have this field and
caused errors. So, we update save_model to be able to accept forms
without a `locales` field.