Fix bug 1091939: Add option to serve pre-generated snippet bundles.
Instead of generating and returning snippet code whenever requested,
return a redirect to a static file with the snippet code. The file is
generated once when requested, and cached for some period of time.

Bundles are grouped by a key generated from the snippets included in the
bundle, as well the client’s locale and about:home version, both of
which affect the snippet code that is sent. This allows reuse of the
snippet bundle across client configurations that would otherwise receive
the exact same snippet code.