Cache bundles for a month.
Bundle filename is calculated based on snippet id and snippet modified date. Thus if contents change the bundle filename changes too. By instructing the browser to cache the bundle for a month we reduce the traffic transferred in the case where bundles do not change between the 4 hour intervals or for a day or two.

We still keep the settings.SNIPPET_BUNDLE_TIMEOUT setting of the fetch_pregenerated_snippets view, which means that the browser will try to get new snippets after settings.SNIPPET_BUNDLE_TIMEOUT and in the case that the bundle hasn't changed it will be redirected to something already cached.