Push Django 1.4.21.
This is a version downgrade from Django 1.5.x to Django 1.4. The 1.5.x serries are not supported anymore and we need to upgrade to 1.8 with LTS support until Oct 2015.

Earlier commits like f48c353 falsely claim the upgrade of the project to Django 1.6 serries which was never done due to the existance of django in both vendor-local and vendor. Let's kill submodules!