Update cache keys when snippet_{js|css} change.
If snippet_js.html or snippet_css.html change but the served snippets do
not change the cache key will not get updated and therefore clients will
still get old content.

In practice this an edge case scenario that we figured out that's
happening due to failing functional tests. The tests always hit a url
that returns no snippets on production and therefore it's serving an old
cached version. The tests then failed because the content was not the
expected one.