[bug 1236090] Wait 500ms before reloading after block.
Waiting for 500ms before reloading the page after user blocks a snippet,
will allow the IndexedDB more time to complete its transaction. While
this is an imperfect non-deterministic solution it will fix the issue
for most of the users.

Ideally we would create our own connection to the iDB and ensure that the
write completes before we reload the page.