stopPropagation while block snippet button gets clicked.
In some snippet templates which can become globally clickable (i.e.
snippet as a button) the block button in included in a global <a>
element. In some cases and FF versoisn when clicking the block button
the event propages all the way to the <a> and gets the user in another
page. stopPropagation fixes this.