Refactor locales.
Snippets are using a OneToOne relation with SnippetLocale. JSONSnippets
act similarly with JSONSNippetLocale.

This commit creates a TargetedLocale model similar to the
TargetedCountry and refactors Snippets and JSONSnippets to store locales
in this ManyToMany relation.

Admin form funkiness gets replaced by Django's built-in functionality,
code is cleaner and will allow easier DB queries in the future.

Data migrations copy current locales to TargetedLocale.

SnippetLocale and JSONSnippetLocale are not deleted yet.