Cache Database queries in LocMem for 5 minutes with django-cachalot.
This will reduce the stress on MariaDB and will not increase the traffic to
Redis in the cost of a little more memory usage per process.

The fact the LocMem is not shared between processes does not raise any issues
except that we use more local memory. DB remains the only source of truth and
the cache will get invalidated every 5 minutes to keep up to date with it.