[Fix #1080] Auto translate only visible fields.
The AutoTranslatorWidget was changing fields in hidden InlineTemplates, making
Django to consider the later `changed` and run validators against them. The
hidden Inline Templates are not supposed to be populated and thus they didn't
have values for the required fields which resulted in Django raising a cryptic
`Fix the errors below` messages, without errors being displayed (b/c the inlines
are hidden).

This commit fixes the issue by having the making AutoTranslatorWidget translate
only visible fields. For a complete solution we need to fix #1082 as well.