[Issue #1112] Redirect all requests to CDN / S3
Redirect all requests to CDN / S3 without checking first if the bundle exists on
S3.

CloudFront is configured to return a custom page (which is an empty JSON file)
when origins return 403 (Permission Denied). S3 bucket will return 403 -and not
404- when you request a file that doesn't exist because it disallows directory
listing.

This way we can return valid 404s from the Django app when needed and empty
bundles when we don't have a snippets for a channel / locale / distribution
combination, without spending time on I/O between the Django App and S3 or
hitting the cache too many times.