[Issue #1413] Fetch DailyImpressions only once per week.
DailyImpressions are used to calculate an approximation of how many
Impressions are actually valid impressions, i.e. User has spend enough
time on the page to read and understand the message.

Metrics have shown that the percentage of time spend of New Tab Page is
pretty stable and around 63% of the time Users stay 5 seconds of more on
the page.

Since the number is pretty stable, we don't need to pull it every day.
Instead this PR changes the pull frequency to once per week to save on
resources.