[Fix #1399] Include total number of clients for Job.
Fetches total number of clients for each Job when Job completes with a
special query and scans data from `publish_start` to `completed_on`.

This limits that amount of data scanned while it maintains
functionality. The only downside is that total number of clients that
interract with the Job (Snippet) is known well the Job finishes
publication instead of while the Job runs.

For Jobs that run for more than 30 days -usually  F100s- we don't
fetch client data to avoid scanning a too big dataset.